Refactor: Remove unused database modules and imports
This commit removes several database-related modules and imports that are no longer necessary. This includes the removal of app/db/__init__.py, app/db/manager.py, app/db/migration_manager.py, app/db/migrations/*, app/db/models.py, app/db/rag_service.py, app/db/scheduler_db_service.py, and app/db/service.py.
