+++ app/services/db/__init__.py
@@ -8,10 +8,10 @@ This module handles all database operations including:
 - Scheduler job persistence
 """
 
-from app.db.manager import DatabaseManager
-from app.db.models import CronJob, JobStatus, Message, ScheduleType
-from app.db.scheduler_db_service import SchedulerDatabaseService
-from app.db.service import DBService
+from app.services.db.manager import DatabaseManager
+from app.services.db.models import CronJob, JobStatus, Message, ScheduleType
+from app.services.db.scheduler_db_service import SchedulerDatabaseService
+from app.services.db.service import DBService
 
 __all__ = [
     "DBService",
