+++ app/services/stt_audio_input/service.py
