+++ app/services/stt_transcription/service.py
