+++ app/services/tooling/mcp/mcp_discovery.py
