+++ app/services/tooling/tools/duckduckgo_search.py
